third_party/blink/web_tests/svg/dom/svgpath-negative-pathLength-expected.txt

CONSOLE ERROR: Error: <path> attribute pathLength: A negative value is not valid. ("-1")
